Restructure placement of edit profile fragment
