+++ app/src/main/res/layout/edit_profile_fragment.xml
@@ -7,7 +7,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/edit_layout"
     android:background="#fff"
-    tools:context=".ui.EditProfileFragment">
+    tools:context=".ui.user_profile.edit.EditProfileFragment">
 
 
         <TextView
